+++ packages/payload-mux/src/lib/mergeCollection/index.ts
@@ -1,0 +1,3 @@
+export { mergeCollection } from './mergeCollection'
+export { mergeHooks } from './mergeHooks'
+export { mergeSelect } from './mergeSelect'